pyproject.toml
[tool.poetry]
name = "admin_db"
version = "0.1.0"
description = ""
authors = [""]

[tool.poetry.dependencies]
python = "^3.8"
aiogram = "^2.20"
python-dotenv = "^0.20.0"
peewee = "^3.15.0"
peewee-migrate = "^1.4.8"
APScheduler = "^3.9.1"
SQLAlchemy = "^1.4.39"
mysqlclient = "^2.1.1"
openai = "^0.27.7"
requests = "^2.31.0"
fastapi = "^0.96.0"
uvicorn = "^0.22.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
